Move some helper functions to a top level internal package
