+++ conflow/schema/boolean.go
@@ -13,6 +13,8 @@ import (
 	"fmt"
 	"reflect"
 	"strconv"
+
+	"github.com/conflowio/conflow/internal/utils"
 )
 
 type Boolean struct {
@@ -25,7 +27,7 @@ type Boolean struct {
 
 func (b *Boolean) AssignValue(imports map[string]string, valueName, resultName string) string {
 	if b.Pointer {
-		schemaPackageName := EnsureUniqueGoPackageName(imports, "github.com/conflowio/conflow/conflow/schema")
+		schemaPackageName := utils.EnsureUniqueGoPackageName(imports, "github.com/conflowio/conflow/conflow/schema")
 		return fmt.Sprintf("%s = %s.BooleanPtr(%s.(bool))", resultName, schemaPackageName, valueName)
 	}
