+++ conflow/schema/number_test.go
@@ -10,12 +10,13 @@ import (
 	"encoding/json"
 	"errors"
 
+	"github.com/conflowio/conflow/internal/testhelper"
+
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/ginkgo/extensions/table"
 	. "github.com/onsi/gomega"
 
 	"github.com/conflowio/conflow/conflow/schema"
-	"github.com/conflowio/conflow/conflow/schema/internal/testhelper"
 )
 
 var _ schema.Schema = &schema.Number{}
